working prototype of idena bindgen

+++ sdk-core/assembly/allocate.ts
@@ -1,7 +1,7 @@
 import {Region} from "./region";
 import {debug} from "./debug";
 
-export function allocate(size: usize): i32 {
+export function allocate(size: u32): usize {
   let data = new Uint8Array(size);
   let r = new Region(data);
   let result = changetype<usize>(r);
